Restructured API Gateway
