--- backend/microservices/nope.env
@@ -1,17 +1,0 @@
-# API Configuration
-API_HOST=localhost
-API_PORT=5001
-
-# CORS Configurationasd
-CORS_ORIGINS=http://localhost:5173,http://localhost:3000,http://localhost:5001
-
-# Redis Configuration
-REDIS_HOST=localhost
-REDIS_PORT=6379
-
-# API Keys
-NEWS_API_KEY=4b94554081e148bc964e4ab94c9dc0fe
-OPENAI_API_KEY=your_openai_api_key_here
-
-# Logging
-LOG_LEVEL=INFO
